cln: refactor plotting
